Large Update
11/26 Early Early Morning: I just finished what seems to be the final
for the start of our Red Box. I changed the code up quite a bit. Major
updates: I changed the TestGUI so that there were no scenes involved.
This allows us to just use Frames and set them to visible or not when we
need them. Also I was able to use a method that allows us to make the
frame full screen for a more professional feel. I changed the StartGUI
so now it has tabs instead of buttons. It seems a little cleaner and
easier to use. The submit button still works the same. Still need to
implement a New User Button. Also this tabs thing might work for
displaying our videos/games.